moved a util out of the middleware and into utils as well as replaced values with enviromental variables
